refactor(nms): Consolidate Text components
Remove text compoment that uses symphony theme and replace its usages.
