+++ nms/app/components/GatewaySummaryFields.js
@@ -27,7 +27,7 @@ import FormField from './FormField';
 import Input from '@material-ui/core/Input';
 import MagmaV1API from '../../generated/WebClient';
 import React from 'react';
-import Text from '../../fbc_js_core/ui/components/design-system/Text';
+import Text from '../theme/design-system/Text';
 import moment from 'moment';
 
 import nullthrows from '../../fbc_js_core/util/nullthrows';
