+++ nms/app/components/VersionTooltip.js
@@ -17,7 +17,7 @@
 import React from 'react';
 
 import AppContext from '../../fbc_js_core/ui/context/AppContext';
-import Text from '../../fbc_js_core/ui/components/design-system/Text';
+import Text from '../theme/design-system/Text';
 import Tooltip from '@material-ui/core/Tooltip';
 import {makeStyles} from '@material-ui/styles';
