+++ nms/app/components/insights/Metrics.js
@@ -26,7 +26,7 @@ import FormControl from '@material-ui/core/FormControl';
 import GridList from '@material-ui/core/GridList';
 import GridListTile from '@material-ui/core/GridListTile';
 import InputLabel from '@material-ui/core/InputLabel';
-import Text from '../../../fbc_js_core/ui/components/design-system/Text';
+import Text from '../../theme/design-system/Text';
 import TextField from '@material-ui/core/TextField';
 import TimeRangeSelector from './TimeRangeSelector';
