+++ nms/app/views/metrics/CloudMetrics.js
@@ -24,7 +24,7 @@ import CardContent from '@material-ui/core/CardContent';
 import GridList from '@material-ui/core/GridList';
 import GridListTile from '@material-ui/core/GridListTile';
 import React from 'react';
-import Text from '../../../fbc_js_core/ui/components/design-system/Text';
+import Text from '../../theme/design-system/Text';
 import TimeRangeSelector from '../../components/insights/TimeRangeSelector';
 
 import {makeStyles} from '@material-ui/styles';
