+++ nms/fbc_js_core/ui/components/design-system/Input/InputAffix.js
@@ -16,7 +16,7 @@
 
 import * as React from 'react';
 import SymphonyTheme from '../../../theme/symphony';
-import Text from '../Text';
+import Text from '../../../../../app/theme/design-system/Text';
 import classNames from 'classnames';
 import {makeStyles} from '@material-ui/styles';
 import {useFormElementContext} from '../Form/FormElementContext';
