+++ nms/fbc_js_core/ui/host/OrganizationUsersTable.js
@@ -15,7 +15,7 @@
  */
 import ActionTable from '../components/ActionTable';
 import React from 'react';
-import Text from '../../../fbc_js_core/ui/components/design-system/Text';
+import Text from '../../../app/theme/design-system/Text';
 import axios from 'axios';
 import withAlert from '../../../fbc_js_core/ui/components/Alert/withAlert';
 import type {EditUser} from './OrganizationEdit';
