+++ nms/fbc_js_core/ui/host/Organizations.js
@@ -37,7 +37,7 @@ import OrganizationDialog from './OrganizationDialog';
 import PersonAdd from '@material-ui/icons/PersonAdd';
 import PersonIcon from '@material-ui/icons/Person';
 import React from 'react';
-import Text from '../components/design-system/Text';
+import Text from '../../../app/theme/design-system/Text';
 import axios from 'axios';
 import withAlert from '../../../fbc_js_core/ui/components/Alert/withAlert';
